3 - Stateless and Stateful
In this class we better understand how to use Material Design in our project, in addition to creating our own custom widget! We now have more control over how to manipulate our application's structures to create a predefined screen. Shall we recap?

Scaffold:
The widget that follows the Material Design rules to create a screen with appBar, body, floatingActionButton, in addition to other parameters that we can check in the Widgets catalog in the documentation.

Stateless Widget:
Pre-prepared structure to create our custom widget. Static screen widget, that is, that does not change with time or actions.

ListView:
Widget that allows showing a list of elements on the screen, with the possibility of dragging the screen. With this Widget we started to discover how to explore screens bigger than our device's space limit. Armed with the parameter children, it also has other parameters that we can check in the Widgets catalog in the documentation.

Padding:
The Widget that allows us to create an outer layer of space on each of the 4 sides of its child Widget (child) using the EdgeInsets parameter padding.

Stateful Widget:
Pre-prepared structure to create our custom widget, a dynamic screen widget, that is, that can change with time or actions.

LinearProgressIndicator:
The first Widget from the catalog we used is already a Stateful by nature! It has the ability to change size according to its value (value) in double. We also have the option to change its background color (backgroundColor) and main color (color), in addition to other parameters that we can check in the Widgets catalog in the documentation.

Hot Reload e Hot Restart:
With the discovery of Stateful we can now understand the big difference between Hot Restart and Hot Reload. We saw how Hot Restart restarts our application by reloading the values and redrawing the screen, without saving any information (takes a little longer), while Hot Reload reloads some simple elements on the screen and keeps the values saved (extremely fast).