4 - Images and animations
Finally we managed to reproduce our Layout! And we managed to make our project work perfectly, in addition to applying small details that make our application more refined . But wait a minute, we need to revisit the subjects to fix everything in the head.

Image:
The Widget has the power to add images to our project: images from the internet, assets, files and even from device memory. It has parameters to control the vertical size (height), the horizontal size (width), the fit of the photo (fit), in addition to other parameters that we can check in the Widgets catalog in the documentation.

BoxDecoration:
To fine-tune our Layout we use the decorationContainer parameter. From there we discovered the power of BoxDecorationto manipulate the edges using the borderRadius. We also noticed that when we use decorationthe parameter coloronly needs to be used within the BoxDecoration. Remember that we can check other parameters in the Widgets catalog in the documentation.

AnimatedOpacity:
Our first Widget with animation , allowing to manipulate the opacity of its child Widget (child) with a pre-established duration (duration).