5 - Good Practices
You got it! You have reached the end of the course, congratulations! And to close with a golden key, let's go to one last step: Looking back and seeing how much you've learned.

Refactoring:
When we finish our project, we have to pay attention to the quality of what we write. Other devs may want to use our project as support, and for that we have to make our code presentable. Because of this, we understand concepts such as single responsibility and the importance of organizing our files into separate folders.

Nomenclature:
In order to make it easy to understand the structure of the code, we have to follow some standards when naming our files, using _spaces instead and using only lowercase letters. In the same way we need to follow the Camel Case pattern for the name of Widgets or variables in our project.

Assets:
We understand the importance of knowing how to choose to use Assets or Network to load images in our project. When using the network, our project is lighter on the device, however the images may take a while to load (or worse not be loaded at all), while using Assets guarantees the fast loading of the images in the project, but makes it heavier on the device than user.

Warnings:
It might just seem TOC, but code with Warnings is inefficient code. We learned to identify the Warnings in our code and remove them one by one, to improve the performance of our project.